requirements.txt
accelerate==1.6.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
mysql-connector-python
anyio==4.9.0
attrs==25.3.0
bitsandbytes==0.42.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.1
cycler==0.12.1
dash==3.0.2
dash-bootstrap-components==2.0.0
dataclasses-json==0.6.7
datasets==3.5.0
defog==0.68.2
dill==0.3.8
diskcache==5.6.3
docker==7.1.0
docopt==0.6.2
docstring_parser==0.16
fastapi==0.115.12
ffmpy==0.5.0
filelock==3.18.0
Flask==3.0.3
fonttools==4.57.0
frozenlist==1.5.0
fsspec==2024.12.0
gitdb==4.0.12
GitPython==3.1.44
gradio==5.24.0
gradio_client==1.8.0
groovy==0.1.2
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.30.2
idna==3.10
importlib_metadata==8.6.1
itsdangerous==2.2.0
Jinja2==3.1.6
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
langchain==0.3.23
langchain-community==0.3.21
langchain-core==0.3.51
langchain-experimental==0.3.4
langchain-ollama==0.3.1
langchain-text-splitters==0.3.8
langsmith==0.3.27
llama_cpp_python==0.3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.1
mdurl==0.1.2
mpmath==1.3.0
multidict==6.2.0
multiprocess==0.70.16
mypy-extensions==1.0.0
mysql==0.0.3
mysql-connector-python==9.2.0
narwhals==1.34.0
nest-asyncio==1.6.0
networkx==3.4.2
numpy==1.26.4
ollama==0.4.7
orjson==3.10.16
packaging==24.2
pandas==2.2.3
peft==0.15.1
pillow==11.1.0
pipreqs==0.4.13
plotly==6.0.1
prompt_toolkit==3.0.50
propcache==0.3.1
protobuf==5.29.4
psutil==7.0.0
psycopg2-binary==2.9.10
pwinput==1.0.3
pyarrow==19.0.1
pydantic==2.11.3
pydantic-settings==2.8.1
pydantic_core==2.33.1
pydeck==0.9.1
pydub==0.25.1
Pygments==2.19.1
PyMySQL==1.1.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
retrying==1.3.4
rich==14.0.0
rpds-py==0.24.0
ruff==0.11.5
safehttpx==0.1.6
safetensors==0.5.3
scipy==1.15.2
semantic-version==2.10.0
setuptools==78.1.0
shellingham==1.5.4
shtab==1.7.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.40
sqlcoder==0.2.5
sqlparse==0.5.3
starlette==0.46.1
streamlit==1.44.1
sympy==1.13.1
tabulate==0.9.0
tenacity==9.1.2
tokenizers==0.21.1
toml==0.10.2
tomlkit==0.13.2
torch==2.6.0
tornado==6.4.2
tqdm==4.67.1
transformers==4.51.1
trl==0.9.6
typeguard==4.4.2
typer==0.15.2
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.1
tyro==0.9.18
tzdata==2025.2
urllib3==2.3.0
uvicorn==0.34.0
wcwidth==0.2.13
websockets==15.0.1
Werkzeug==3.0.6
xxhash==3.5.0
yarg==0.1.10
yarl==1.19.0
zipp==3.21.0
zstandard==0.23.0
